Rename handler.py to file.py
